editor/libeditor/tests/test_bug757771.html

<!DOCTYPE html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=757771
-->
<title>Test for Bug 757771</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=757771">Mozilla Bug 757771</a>
<input value=foo maxlength=4>
<input type=password value=password>
<script>
/** Test for Bug 757771 **/

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  var textInput = document.querySelector("input");
  textInput.focus();
  textInput.select();
  sendString("abcde");

  var passwordInput = document.querySelector("input + input");
  passwordInput.focus();
  passwordInput.select();
  sendString("hunter2");

  ok(true, "No real tests, just crashes/asserts");

  SimpleTest.finish();
});
</script>
